Rename command, clean up files
