+++ Source/SwiftLintFramework/String+SwiftLint.swift
@@ -26,6 +26,10 @@ extension String {
     func countOfTailingCharactersInSet(characterSet: NSCharacterSet) -> Int {
         return String(reverse(self)).countOfLeadingCharactersInSet(characterSet)
     }
+
+    public var chomped: String {
+        return self.stringByTrimmingCharactersInSet(NSCharacterSet.newlineCharacterSet())
+    }
 }
 
 extension NSString {
